create gifs (visualtests are not fixed yet)
