+++ CMakeLists.txt
@@ -21,6 +21,7 @@ endif()
 include(Deps.cmake)
 add_subdirectory(MDLIB)
 add_subdirectory(SETS)
+add_subdirectory(GnuPlot)
 enable_testing()
 add_subdirectory(TESTS)
 if (VIS)
