+++ VISUAL_TESTS/RiftingChenin.txt
@@ -7,7 +7,8 @@ irestart = 0
 writer = 1
 writer_step = 50
 writer_markers = 0
-writer_debug   = 0 
+writer_debug   = 0
+writerSubfolder = RiftingChenin
 
 /**** INPUT FILE FOR PARTICLES ****/
 input_file = setup_r501x501_fsp40.bin
