+++ misc/Python_visualisation/Read_Doodz_Phases.py
@@ -2,7 +2,7 @@ import h5py as h5py
 import numpy as np
 import matplotlib.pyplot as mpl
 
-file   = h5py.File('/home/roman/CLionProjects/MDOODZ7.0/cmake-exec/CollisionIra/result6/Output00200.gzip.h5', 'r')
+file   = h5py.File('/home/roman/CLionProjects/MDOODZ7.0/cmake-exec/CollisionIra/heavy3/Output01740.gzip.h5', 'r')
 phases = file['/VizGrid/compo_hr']
 xv_ph  = file['/VizGrid/xviz_hr']
 zv_ph  = file['/VizGrid/zviz_hr']
